+++ PGLib/src/ParallaxGenPlugin.cpp
@@ -367,17 +367,16 @@ void ParallaxGenPlugin::libSetModelRecNIF(const int& modelRecHandle, const wstri
 }
 
 // Statics
-unordered_map<int, unsigned int> ParallaxGenPlugin::s_txstFormIDs;
+unordered_map<string, unsigned int> ParallaxGenPlugin::s_txstFormIDs;
+unordered_map<string, unsigned int> ParallaxGenPlugin::s_newTXSTFormIDs;
 unordered_set<unsigned int> ParallaxGenPlugin::s_txstResrvedFormIDs;
+unordered_set<unsigned int> ParallaxGenPlugin::s_txstUsedFormIDs;
 unsigned int ParallaxGenPlugin::s_curTXSTFormID = 0;
 
 mutex ParallaxGenPlugin::s_createdTXSTMutex;
 unordered_map<array<wstring, NUM_TEXTURE_SLOTS>, pair<int, string>, ParallaxGenPlugin::ArrayHash,
     ParallaxGenPlugin::ArrayEqual>
     ParallaxGenPlugin::s_createdTXSTs;
-
-mutex ParallaxGenPlugin::s_edidCounterMutex;
-int ParallaxGenPlugin::s_edidCounter = 0;
 
 ParallaxGenDirectory* ParallaxGenPlugin::s_pgd;
 
@@ -408,10 +407,12 @@ void ParallaxGenPlugin::initialize(const BethesdaGame& game, const filesystem::p
 
 void ParallaxGenPlugin::loadTXSTCache(const nlohmann::json& txstCache)
 {
+    s_newTXSTFormIDs.clear();
+
     // loop through json objects
     for (const auto& [key, value] : txstCache.items()) {
         // convert key to array
-        s_txstFormIDs[stoi(key)] = value.get<unsigned int>();
+        s_txstFormIDs[key] = value.get<unsigned int>();
         s_txstResrvedFormIDs.insert(value.get<unsigned int>());
     }
 }
@@ -421,8 +422,8 @@ auto ParallaxGenPlugin::getTXSTCache() -> nlohmann::json
     nlohmann::json txstCache;
 
     // loop through map
-    for (const auto& [key, value] : s_txstFormIDs) {
-        txstCache[to_string(key)] = value;
+    for (const auto& [key, value] : s_newTXSTFormIDs) {
+        txstCache[key] = value;
     }
 
     return txstCache;
@@ -451,9 +452,13 @@ void ParallaxGenPlugin::processShape(const wstring& nifPath, nifly::NiShape* nif
         string altTexJSONKey;
         string txstJSONKey;
 
+        const auto altTexFormIDTuple = libGetAltTexFormID(altTexIndex);
+        const auto txstFormIDCacheKey = ParallaxGenUtil::utf16toUTF8(get<1>(altTexFormIDTuple)) + "/"
+            + to_string(get<0>(altTexFormIDTuple)) + "/" + matchType + "/" + to_string(index3D);
+
         if (PGDiag::isEnabled()) {
             // this is somewhat costly so we only run it if diagnostics are enabled
-            altTexJSONKey = getKeyFromFormID(libGetAltTexFormID(altTexIndex)) + " / " + matchType;
+            altTexJSONKey = getKeyFromFormID(altTexFormIDTuple) + " / " + matchType;
             txstJSONKey = getKeyFromFormID(libGetTXSTFormID(txstIndex));
         }
 
@@ -630,22 +635,36 @@ void ParallaxGenPlugin::processShape(const wstring& nifPath, nifly::NiShape* nif
 
             // Create a new TXST record
             spdlog::trace(L"Plugin Patching | {} | {} | Creating a new TXST record and patching", nifPath, index3D);
-            const string newEDID = fmt::format("PGTXST{:05d}", s_edidCounter++);
 
-            // check if original TXST id exists in reserved IDs
-            if (s_txstFormIDs.contains(txstIndex)) {
+            // Find formID to use
+            unsigned int newFormID = 0;
+            if (s_txstFormIDs.contains(txstFormIDCacheKey)
+                && !s_txstUsedFormIDs.contains(s_txstFormIDs[txstFormIDCacheKey])) {
                 // use old formid for new record
-                curResult.txstIndex = libCreateNewTXSTPatch(altTexIndex, newSlots, newEDID, s_txstFormIDs[txstIndex]);
+                newFormID = s_txstFormIDs[txstFormIDCacheKey];
             } else {
                 // find next available formid
-                unsigned int newFormID = ++s_curTXSTFormID;
-                while (s_txstResrvedFormIDs.contains(newFormID)) {
-                    ++newFormID;
-                }
-
-                curResult.txstIndex = libCreateNewTXSTPatch(altTexIndex, newSlots, newEDID, newFormID);
-                s_txstFormIDs[txstIndex] = newFormID;
+                while (s_txstResrvedFormIDs.contains(++s_curTXSTFormID)) { }
+                newFormID = s_curTXSTFormID;
             }
+
+            if (newFormID == 0) {
+                throw runtime_error("Failed to find a new form ID for TXST record");
+            }
+
+            if (s_txstUsedFormIDs.contains(newFormID)) {
+                throw runtime_error("Form ID already in use");
+            }
+
+            // Insert used FORMID
+            s_txstUsedFormIDs.insert(newFormID);
+
+            // Find EDID
+            const string newEDID = fmt::format("PGTXST{:06X}", newFormID);
+
+            // create new TXST record with chosen form ID
+            curResult.txstIndex = libCreateNewTXSTPatch(altTexIndex, newSlots, newEDID, newFormID);
+            s_newTXSTFormIDs[txstFormIDCacheKey] = newFormID;
 
             patchers.shaderPatchers.at(winningShaderMatch.shader)
                 ->processNewTXSTRecord(winningShaderMatch.match, newEDID);