+++ PGPatcher/include/ParallaxGenConfig.hpp
@@ -34,7 +34,7 @@ public:
             ModManagerDirectory::ModManagerType type = ModManagerDirectory::ModManagerType::NONE;
             std::filesystem::path mo2InstanceDir;
             std::wstring mo2Profile = L"Default";
-            bool mo2UseOrder = true;
+            bool mo2UseOrder = false;
 
             auto operator==(const ModManager& other) const -> bool
             {